changed MyWebViewClient and MyWebChromeClient to single classes
