+++ app/src/main/java/com/katahiromz/krakra_ja_jp/MyWebViewClient.kt
@@ -1,0 +1,19 @@
+package com.katahiromz.krakra_ja_jp
+
+import android.webkit.WebResourceRequest
+import android.webkit.WebView
+import android.webkit.WebViewClient
+
+class MyWebViewClient(activity: MainActivity) : WebViewClient() {
+
+    override fun shouldOverrideUrlLoading(
+        view: WebView?,
+        request: WebResourceRequest?
+    ): Boolean {
+        return false
+    }
+
+    override fun onPageFinished(view: WebView?, url: String?) {
+        super.onPageFinished(view, url)
+    }
+}
